+++ packages/af-utils-react-virtual-headless/src/constants/index.js
@@ -9,9 +9,8 @@ export const SIZES_HASH_MODULO = 2 ** 30 - 1;
 */
 export const MAX_ITEM_COUNT = 0x7fffffff;
 
-export const EVT_FROM = 0;
-export const EVT_TO = 1;
-export const EVT_SCROLL_SIZE = 2;
-export const EVT_SIZES = 3;
+export const EVT_RANGE = 0;
+export const EVT_SCROLL_SIZE = 1;
+export const EVT_SIZES = 2;
 
-export const EVT_ALL = [EVT_FROM, EVT_TO, EVT_SCROLL_SIZE, EVT_SIZES];
+export const EVT_ALL = [EVT_RANGE, EVT_SCROLL_SIZE, EVT_SIZES];