+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-virtual-list
 
+## 0.0.34
+
+### Patch Changes
+
+-   restructured events
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.32
+
 ## 0.0.33
 
 ### Patch Changes